Refactor ServiceManager to MetadataManager that manages loading and validating metadata and ControlService that implements control service and message routing
