+++ npc_engine/server/utils.py
