+++ npc_engine/service_clients/control_client.py
@@ -60,3 +60,8 @@ class ControlClient(ServiceClient):
             "params": [],
         }
         return self.send_request(request)
+
+    @classmethod
+    def get_api_name(cls) -> str:
+        """Return the name of the API."""
+        return "control"
