+++ npc_engine/service_clients/similarity_client.py
@@ -27,3 +27,8 @@ class SimilarityClient(ServiceClient):
         }
         reply = self.send_request(request)
         return reply
+
+    @classmethod
+    def get_api_name(cls) -> str:
+        """Return the name of the API."""
+        return "SimilarityAPI"
