+++ tests/integration/test_run_apis.py
@@ -21,9 +21,9 @@ class TestClass:
                 "python",
                 cli_path,
                 "--verbose",
-                "run",
                 "--models-path",
                 models_path,
+                "run",
                 "--port",
                 "5555",
                 "--start-all",
